Fixed: ui

+++ front/src/services/api.jsx
@@ -35,7 +35,10 @@ api.interceptors.response.use(
       return Promise.reject(error);
     }
 
-    if (error?.response?.status === 401) {
+    if (
+      error?.response?.status === 401 &&
+      !originalConfig?.url?.includes("/auth/sign-in")
+    ) {
       TokenService.removeTokens();
       window.location.href = "/login";
       return Promise.reject("Session expired. Redirecting to login.");
